pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qa-explorer"
version = "0.1.0"
description = "AI-powered exploratory testing agent — give it a URL, it finds bugs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.44.0",
    "pyyaml>=6.0",
    "anthropic>=0.30.0",
    "openai>=1.30.0",
]

[project.scripts]
qa-explorer = "main:main"

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[tool.setuptools]
py-modules = ["main", "config"]

[tool.setuptools.packages.find]
include = ["agent*", "browser*", "models*"]
